Changed the name of some containers and will restructure the project files; Finished the timeline structure, still have to style the TimelineItemDescription;
