+++ src/components/containers/MainInformationSection/AboutContainer/AboutContainer.styled.tsx
