+++ src/components/containers/MainInformationSection/MainInformationContainer.styled.tsx
