+++ src/components/images/Images.styled.tsx
@@ -12,10 +12,8 @@ export const ExperienceImage = styled(Image)`
     border: ${({ theme }) =>
         `${theme.border.medium} solid ${theme.colors.surface.secondary}`};
     border-radius: 50%;
-    transition: transform 0.2s, border 0.2s;
+    transition: transform 0.2s;
     &:hover {
-        transform: scale(1.2);
-        border: ${({ theme }) =>
-            `${theme.border.large} solid ${theme.colors.surface.secondary}`};
+        transform: scale(1.15);
     }
 `;
